CHange return type of fileService.fileDownload (original: Pair, now: ResponseEntity)
