+++ src/main/kotlin/com/navi/server/web/FileApiController.kt
@@ -46,24 +46,6 @@ class FileApiController (val fileService: FileService){
 
     @GetMapping("api/navi/files/{token}")
     fun fileDownload(@PathVariable token: String) : ResponseEntity<Resource> {
-        val pair : Pair<FileResponseDTO, Resource> = fileService.fileDownload(token) ?: run {
-            throw NotFoundException("File Not Found")
-        }
-        val fileResponseDTO: FileResponseDTO = pair.first.apply {
-            val osType: String = System.getProperty("os.name").toLowerCase()
-
-            if (osType.indexOf("win") >= 0) {
-                this.fileName = this.fileName.split("\\").last()
-            } else {
-                this.fileName = this.fileName.split("/").last()
-            }
-        }
-
-        val again: String = String.format("attachment; filename=\"%s\"", URLEncoder.encode(fileResponseDTO.fileName, "UTF-8"))
-        val resource: Resource = pair.second
-        return ResponseEntity.ok()
-                .contentType(MediaType.parseMediaType("application/octet-stream"))
-                .header(HttpHeaders.CONTENT_DISPOSITION, again)
-                .body(resource)
+        return fileService.fileDownload(token)
     }
 }
