+++ src/test/kotlin/com/navi/server/service/FileServiceTest.kt
@@ -13,6 +13,8 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.core.io.Resource
+import org.springframework.http.HttpHeaders
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.mock.web.MockMultipartFile
@@ -320,7 +322,6 @@ class FileServiceTest {
 
     @Test
     fun fileDownloadTest() {
-
         // Make one test file to root
         val fileName: String = "downloadTest-service.txt"
         val fileObject: File = File(fileConfigurationComponent.serverRoot, fileName)
@@ -334,16 +335,16 @@ class FileServiceTest {
 
         // file Download
         val targetToken = fileService.getSHA256(fileObject.absolutePath)
-        val result = fileService.fileDownload(targetToken)
+        val responseEntity : ResponseEntity<Resource> = fileService.fileDownload(targetToken)
 
         // Assert
-        val fileResponseDTO = result?.first
-        val resource = result?.second
+        val contentDisposition : String? = responseEntity.headers.get(HttpHeaders.CONTENT_DISPOSITION)!!.get(0)
+        contentDisposition?.let {
+            val resultFileName = it.split("=")[1]
+            assertThat(resultFileName.substring(1, resultFileName.length-1)).isEqualTo(fileName)
+        } ?: throw Exception("File Name mismatch OR No File Name in ContentDisposition")
 
-        fileResponseDTO?.let {
-            assertThat(fileResponseDTO.fileName).isEqualTo(fileObject.absolutePath)
-        } ?: throw Exception("No File: ${fileObject.absolutePath}")
-
+        val resource : Resource? = responseEntity.body
         resource?.let {
             val resultContent = resource.inputStream.readBytes().toString(Charsets.UTF_8)
             assertThat(resultContent).isEqualTo(fileContent)
